+++ R/yprBH_slot_func.R
@@ -154,9 +154,9 @@ yprBH_slot_func <- function(lowerSL,upperSL,cfBelow,cfIn,cfAbove,cm,lhparms,
     avglen_in <- 10^((log10(avgwt_in) - LWalpha)/LWbeta)
   }
 
-  #yield over slot######
+  #yield Above slot######
 
-  #Parameters for over slot
+  #Parameters for Above slot
   F_Above <- -1*log(1-cfAbove)
   M_Above <- -1*log(1-cm)
   Z_Above <- F_Above+M_Above
@@ -260,7 +260,7 @@ yprBH_slot_func <- function(lowerSL,upperSL,cfBelow,cfIn,cfAbove,cm,lhparms,
     nrAbove=Nr_Above,
     trBelow=tr,
     trIn=tmax_lowerSL,
-    trOver=tmax_upperSL,
+    trAbove=tmax_upperSL,
     avglenBelow=avglen_Below,
     avglenIn=avglen_in,
     avglenAbove=avglen_Above,
@@ -287,7 +287,7 @@ yprBH_slot_func <- function(lowerSL,upperSL,cfBelow,cfIn,cfAbove,cm,lhparms,
     SAbove=S_Above,
     cfBelow=cfBelow,
     cfIn=cfIn,
-    cfOver=cfAbove,
+    cfAbove=cfAbove,
     recruitmentTL=recruitmentTL,
     lowerSL=lowerSL,
     upperSL=upperSL,